refactor(backend): split ModelController by request pathway
Probing a model and asking a provider what it offers moved out of
ModelController into ModelTestController and ModelDiscoveryController.
What stays is the record list and the state toggles — the module's own
screen — while the two collaborators own the outbound calls, which is
where the provider adapters, the discovery service and the test-prompt
resolver were needed and nowhere else.

The three AJAX routes keep their identifiers and their paths, so no
backend template, no JavaScript and no consumer route lookup changes;
only the class behind the route does. Each new controller carries the
same admin guard as before (ADR-037), now covered by a functional test
per route rather than by proximity to the guarded siblings.

The unit tests moved with the code and the functional and end-to-end
tests were retargeted at the new controllers, so the suite covers the
same behaviour it did before: 5666 unit tests, unchanged.

Same split by request pathway as ADR-027.
