+++ Classes/Controller/Backend/ModelController.php
@@ -14,20 +14,12 @@ use Doctrine\DBAL\Exception as DbalException;
 use Netresearch\NrLlm\Controller\Backend\Response\ErrorResponse;
 use Netresearch\NrLlm\Controller\Backend\Response\ModelListResponse;
 use Netresearch\NrLlm\Controller\Backend\Response\SuccessResponse;
-use Netresearch\NrLlm\Controller\Backend\Response\TestConnectionResponse;
 use Netresearch\NrLlm\Controller\Backend\Response\ToggleActiveResponse;
-use Netresearch\NrLlm\Domain\Enum\ModelCapability;
 use Netresearch\NrLlm\Domain\Model\Model;
 use Netresearch\NrLlm\Domain\Model\Provider;
 use Netresearch\NrLlm\Domain\Repository\ModelRepository;
 use Netresearch\NrLlm\Domain\Repository\ProviderRepository;
-use Netresearch\NrLlm\Provider\Contract\ProviderInterface;
-use Netresearch\NrLlm\Provider\Exception\ProviderException;
-use Netresearch\NrLlm\Provider\ProviderAdapterRegistryInterface;
 use Netresearch\NrLlm\Service\Analytics\AnalyticsPeriod;
-use Netresearch\NrLlm\Service\SetupWizard\DTO\DetectedProvider;
-use Netresearch\NrLlm\Service\SetupWizard\ModelDiscoveryInterface;
-use Netresearch\NrLlm\Service\TestPromptResolverInterface;
 use Netresearch\NrLlm\Service\UsageAnalyticsService;
 use Netresearch\NrLlm\Service\UsageAnalyticsServiceInterface;
 use Psr\Http\Message\ResponseInterface;
@@ -76,9 +68,6 @@ final class ModelController extends ActionController
         private readonly PersistenceManagerInterface $persistenceManager,
         private readonly PageRenderer $pageRenderer,
         private readonly BackendUriBuilder $backendUriBuilder,
-        private readonly ProviderAdapterRegistryInterface $providerAdapterRegistry,
-        private readonly ModelDiscoveryInterface $modelDiscovery,
-        private readonly TestPromptResolverInterface $testPromptResolver,
         private readonly UsageAnalyticsServiceInterface $analytics,
         private readonly LoggerInterface $logger,
     ) {}
@@ -257,356 +246,6 @@ final class ModelController extends ActionController
     }
 
     /**
-     * AJAX: Test model by making a simple API call.
-     */
-    public function testModelAction(ServerRequestInterface $request): ResponseInterface
-    {
-        if (($deny = $this->denyNonAdmin()) !== null) {
-            return $deny;
-        }
-        $body = $request->getParsedBody();
-        $uid = $this->extractIntFromBody($body, 'uid');
-
-        $model = $this->resolveModelOrError($uid);
-        if (!$model instanceof Model) {
-            return $model;
-        }
-
-        return $this->performModelTest($model, $uid);
-    }
-
-    /** Send a chat prompt through the record's own provider. */
-    private const PROBE_CHAT = 'chat';
-
-    /** Embed a string through the record's own provider. */
-    private const PROBE_EMBEDDINGS = 'embeddings';
-
-    /** Nothing here can verify this model — say so rather than guess. */
-    private const PROBE_UNSUPPORTED = 'unsupported';
-
-    /**
-     * Capabilities a chat completion exercises. Any of them means the model
-     * answers a prompt, which is what the chat probe sends.
-     */
-    private const CHAT_SHAPED_CAPABILITIES = [
-        ModelCapability::CHAT,
-        ModelCapability::COMPLETION,
-        ModelCapability::VISION,
-        ModelCapability::TOOLS,
-        ModelCapability::STREAMING,
-        ModelCapability::JSON_MODE,
-        // Audio is a chat-completions modality in every provider this
-        // extension speaks to — there is no audio service and no separate
-        // credential, so a record declaring only it still answers a prompt.
-        ModelCapability::AUDIO,
-    ];
-
-    /**
-     * Perform the actual provider test call for a resolved model.
-     */
-    private function performModelTest(Model $model, int $uid): ResponseInterface
-    {
-        // Provider is lazy-loaded by Extbase
-        if ($model->getProvider() === null) {
-            return new JsonResponse((new ErrorResponse($this->localize('LLL:EXT:nr_llm/Resources/Private/Language/locallang.xlf:error.model.noProviderConfigured', 'Model has no provider configured')))->jsonSerialize(), 400);
-        }
-
-        // Resolve localized format strings before try/catch to avoid LocalizationUtility
-        // exceptions being caught by the adapter error handler below
-        $respondedFormat = $this->translate('model.test.responded')
-            ?? 'Model "%s" responded: "%s" (tokens: %d in, %d out)';
-        $emptyFormat = $this->translate('model.test.emptyResponse')
-            ?? 'Model "%s" connected successfully (tokens: %d in, %d out) - response content empty, model may use internal reasoning';
-
-        // A model that cannot answer a chat prompt must not be sent one: the
-        // provider rejects it and the admin sees "provider rejected the test
-        // request", which says nothing about the model. Decide from the
-        // record's declared capabilities what a meaningful test even is.
-        $probe = $this->testProbeFor($model);
-        if ($probe === self::PROBE_UNSUPPORTED) {
-            return new JsonResponse((new TestConnectionResponse(
-                success: false,
-                message: sprintf(
-                    $this->translate('model.test.notTestableHere')
-                        ?? 'Model "%s" provides %s. Those services take their credential from the Extension Configuration rather than from this provider record, so this record cannot verify them.',
-                    $model->getName(),
-                    implode(', ', $model->getCapabilitySet()->toStringList()),
-                ),
-            ))->jsonSerialize());
-        }
-
-        try {
-            // Create adapter from model's provider
-            $adapter = $this->providerAdapterRegistry->createAdapterFromModel($model);
-
-            if ($probe === self::PROBE_EMBEDDINGS) {
-                return $this->respondToEmbeddingProbe($model, $adapter);
-            }
-
-            // Make a simple test call - use enough tokens for models with thinking
-            $testPrompt = $this->testPromptResolver->resolve();
-            $response = $adapter->complete($testPrompt, [
-                'model' => $model->getModelId(),
-                'max_tokens' => 100,
-            ]);
-
-            $responseText = trim($response->content);
-
-            // Build success message
-            if ($responseText !== '') {
-                $message = sprintf(
-                    $respondedFormat,
-                    $model->getName(),
-                    mb_substr($responseText, 0, 100) . (mb_strlen($responseText) > 100 ? '...' : ''),
-                    $response->usage->promptTokens,
-                    $response->usage->completionTokens,
-                );
-            } else {
-                // Model connected but returned empty content (might be using thinking mode)
-                $message = sprintf(
-                    $emptyFormat,
-                    $model->getName(),
-                    $response->usage->promptTokens,
-                    $response->usage->completionTokens,
-                );
-            }
-
-            return new JsonResponse((new TestConnectionResponse(
-                success: true,
-                message: $message,
-            ))->jsonSerialize());
-        } catch (ProviderException $e) {
-            // REC #8b: provider error text often references upstream
-            // bodies / endpoints / auth artefacts — log full detail,
-            // surface a short upstream-error message stripped of
-            // internal context. The frontend renders this verbatim
-            // in the test-connection toast.
-            $this->logger->warning('Model test: provider rejected request', [
-                'exception' => $e,
-                'model_uid' => $uid,
-            ]);
-            $message = 'LLM provider rejected the test request. See system log for details.';
-        } catch (Throwable $e) {
-            $this->logger->error('Model test: unexpected error', [
-                'exception' => $e,
-                'model_uid' => $uid,
-            ]);
-            $message = 'Test failed. See system log for details.';
-        }
-
-        return new JsonResponse((new TestConnectionResponse(
-            success: false,
-            message: $message,
-        ))->jsonSerialize());
-    }
-
-    /**
-     * AJAX: Fetch available model IDs from provider's API.
-     *
-     * Uses the ModelDiscoveryService to query the provider's API and return
-     * a list of available models that can be used in the model_id field.
-     */
-    public function fetchAvailableModelsAction(ServerRequestInterface $request): ResponseInterface
-    {
-        if (($deny = $this->denyNonAdmin()) !== null) {
-            return $deny;
-        }
-        $body = $request->getParsedBody();
-        $providerUid = $this->extractIntFromBody($body, 'providerUid');
-
-        if ($providerUid === 0) {
-            return new JsonResponse([
-                'success' => false,
-                'error' => $this->localize('LLL:EXT:nr_llm/Resources/Private/Language/locallang.xlf:error.model.noProviderUid', self::ERROR_NO_PROVIDER_UID),
-            ], 400);
-        }
-
-        $provider = $this->providerRepository->findByUid($providerUid);
-        if ($provider === null) {
-            return new JsonResponse([
-                'success' => false,
-                'error' => $this->localize('LLL:EXT:nr_llm/Resources/Private/Language/locallang.xlf:error.provider.notFound', 'Provider not found'),
-            ], 404);
-        }
-
-        return $this->discoverAvailableModels($provider);
-    }
-
-    /**
-     * Query the provider's API for the list of available models.
-     */
-    private function discoverAvailableModels(Provider $provider): ResponseInterface
-    {
-        try {
-            // Create a DetectedProvider to use with ModelDiscoveryService
-            $detected = new DetectedProvider(
-                adapterType: $provider->getAdapterType(),
-                suggestedName: $provider->getName(),
-                endpoint: $provider->getEffectiveEndpointUrl(),
-            );
-
-            // Discover available models from the provider's API
-            $discoveredModels = $this->modelDiscovery->discover(
-                $detected,
-                $provider->getDecryptedApiKey(),
-            );
-
-            // Convert to simple array for JSON response
-            $models = [];
-            foreach ($discoveredModels as $model) {
-                $models[] = [
-                    'id' => $model->modelId,
-                    'name' => $model->name,
-                    'description' => $model->description,
-                    'contextLength' => $model->contextLength,
-                    'maxOutputTokens' => $model->maxOutputTokens,
-                    'capabilities' => $model->capabilities,
-                    'costInput' => $model->costInput,
-                    'costOutput' => $model->costOutput,
-                    'recommended' => $model->recommended,
-                ];
-            }
-
-            return new JsonResponse([
-                'success' => true,
-                'models' => $models,
-                'providerName' => $provider->getName(),
-                'source' => $this->modelDiscovery->wasLastDiscoveryFromFallback() ? 'fallback' : 'live',
-            ]);
-        } catch (ProviderException $e) {
-            $this->logger->warning('Model fetchAvailableModels: provider error', ['exception' => $e]);
-            $error = 'LLM provider error while fetching model IDs. See system log for details.';
-            $status = 502;
-        } catch (Throwable $e) {
-            $this->logger->error('Model fetchAvailableModels: unexpected error', ['exception' => $e]);
-            $error = 'Failed to fetch model IDs. See system log for details.';
-            $status = 500;
-        }
-
-        return new JsonResponse([
-            'success' => false,
-            'error' => $error,
-        ], $status);
-    }
-
-    /**
-     * AJAX: Detect model limits by querying the provider's API.
-     *
-     * Takes a provider UID and model ID, queries the provider's API,
-     * and returns the model's context length, max output tokens, and capabilities.
-     */
-    public function detectLimitsAction(ServerRequestInterface $request): ResponseInterface
-    {
-        if (($deny = $this->denyNonAdmin()) !== null) {
-            return $deny;
-        }
-        $body = $request->getParsedBody();
-        $providerUid = $this->extractIntFromBody($body, 'providerUid');
-        $modelId = $this->extractStringFromBody($body, 'modelId');
-
-        $validationError = $this->validateDetectLimitsInput($providerUid, $modelId);
-        if ($validationError !== null) {
-            return $validationError;
-        }
-
-        $provider = $this->providerRepository->findByUid($providerUid);
-        if ($provider === null) {
-            return new JsonResponse([
-                'success' => false,
-                'error' => $this->localize('LLL:EXT:nr_llm/Resources/Private/Language/locallang.xlf:error.provider.notFound', 'Provider not found'),
-            ], 404);
-        }
-
-        return $this->discoverModelLimits($provider, $modelId);
-    }
-
-    /**
-     * Validate the detect-limits request input.
-     *
-     * Returns a JSON error response when invalid, or null when the input is valid.
-     */
-    private function validateDetectLimitsInput(int $providerUid, string $modelId): ?ResponseInterface
-    {
-        if ($providerUid === 0) {
-            return new JsonResponse([
-                'success' => false,
-                'error' => $this->localize('LLL:EXT:nr_llm/Resources/Private/Language/locallang.xlf:error.model.noProviderUid', self::ERROR_NO_PROVIDER_UID),
-            ], 400);
-        }
-
-        if ($modelId === '') {
-            return new JsonResponse([
-                'success' => false,
-                'error' => $this->localize('LLL:EXT:nr_llm/Resources/Private/Language/locallang.xlf:error.model.noModelId', 'No model ID specified'),
-            ], 400);
-        }
-
-        return null;
-    }
-
-    /**
-     * Query the provider's API and return the limits for a specific model.
-     */
-    private function discoverModelLimits(Provider $provider, string $modelId): ResponseInterface
-    {
-        try {
-            // Create a DetectedProvider to use with ModelDiscovery
-            $detected = new DetectedProvider(
-                adapterType: $provider->getAdapterType(),
-                suggestedName: $provider->getName(),
-                endpoint: $provider->getEffectiveEndpointUrl(),
-            );
-
-            // Discover available models from the provider's API
-            $discoveredModels = $this->modelDiscovery->discover(
-                $detected,
-                $provider->getDecryptedApiKey(),
-            );
-
-            // Find the specific model
-            $foundModel = null;
-            foreach ($discoveredModels as $model) {
-                if ($model->modelId === $modelId) {
-                    $foundModel = $model;
-                    break;
-                }
-            }
-
-            if ($foundModel === null) {
-                return new JsonResponse([
-                    'success' => false,
-                    'error' => sprintf($this->localize('LLL:EXT:nr_llm/Resources/Private/Language/locallang.xlf:error.model.modelIdNotFound', 'Model "%s" not found in provider\'s available models'), $modelId),
-                ], 404);
-            }
-
-            return new JsonResponse([
-                'success' => true,
-                'modelId' => $foundModel->modelId,
-                'name' => $foundModel->name,
-                'description' => $foundModel->description,
-                'contextLength' => $foundModel->contextLength,
-                'maxOutputTokens' => $foundModel->maxOutputTokens,
-                'capabilities' => $foundModel->capabilities,
-                'costInput' => $foundModel->costInput,
-                'costOutput' => $foundModel->costOutput,
-            ]);
-        } catch (ProviderException $e) {
-            $this->logger->warning('Model detectLimits: provider error', ['exception' => $e]);
-            $error = 'LLM provider error while detecting model limits. See system log for details.';
-            $status = 502;
-        } catch (Throwable $e) {
-            $this->logger->error('Model detectLimits: unexpected error', ['exception' => $e]);
-            $error = 'Failed to detect model limits. See system log for details.';
-            $status = 500;
-        }
-
-        return new JsonResponse([
-            'success' => false,
-            'error' => $error,
-        ], $status);
-    }
-
-    /**
      * Build FormEngine edit URL for a model.
      */
     private function buildEditUrl(int $uid): string
@@ -667,102 +306,4 @@ final class ModelController extends ActionController
         return is_numeric($value) ? (int)$value : 0;
     }
 
-    /**
-     * Extract string value from request body.
-     */
-    private function extractStringFromBody(mixed $body, string $key): string
-    {
-        if (!is_array($body)) {
-            return '';
-        }
-
-        $value = $body[$key] ?? '';
-
-        return is_scalar($value) ? trim((string)$value) : '';
-    }
-
-    private function translate(string $key): ?string
-    {
-        try {
-            return LocalizationUtility::translate(
-                'LLL:EXT:nr_llm/Resources/Private/Language/locallang.xlf:' . $key,
-                'NrLlm',
-            );
-        } catch (Throwable) {
-            return null;
-        }
-    }
-
-    /**
-     * Which probe verifies this model from a provider record.
-     *
-     * Only capabilities served by the record's own provider connection can be
-     * verified here. Image, speech and transcription go through the specialized
-     * services, whose credential comes from the Extension Configuration and not
-     * from this provider — testing them from here would authenticate with a
-     * different secret than the record declares, which is worse than not
-     * testing at all.
-     *
-     * An empty capability list keeps the chat probe: the TCA field has no
-     * minimum, so it is routinely left as shipped, and refusing to test the
-     * common case would be a regression.
-     */
-    private function testProbeFor(Model $model): string
-    {
-        $capabilities = $model->getCapabilitySet();
-        if ($capabilities->isEmpty()) {
-            return self::PROBE_CHAT;
-        }
-
-        foreach (self::CHAT_SHAPED_CAPABILITIES as $capability) {
-            if ($capabilities->has($capability)) {
-                return self::PROBE_CHAT;
-            }
-        }
-
-        if ($capabilities->has(ModelCapability::EMBEDDINGS)) {
-            return self::PROBE_EMBEDDINGS;
-        }
-
-        return self::PROBE_UNSUPPORTED;
-    }
-
-    /**
-     * Embed a short string and report the vector's dimensions.
-     *
-     * The one non-chat capability that runs on the record's own provider
-     * connection, so it verifies exactly what the record claims.
-     */
-    private function respondToEmbeddingProbe(Model $model, ProviderInterface $adapter): ResponseInterface
-    {
-        $response = $adapter->embeddings($this->testPromptResolver->resolve(), [
-            'model' => $model->getModelId(),
-        ]);
-
-        $dimensions = $response->getDimensions();
-        if ($dimensions === 0) {
-            // A 2xx whose body carries no vector has verified nothing; saying
-            // "returned a 0-dimension embedding" in the green panel would be a
-            // false positive.
-            return new JsonResponse((new TestConnectionResponse(
-                success: false,
-                message: sprintf(
-                    $this->translate('model.test.emptyEmbedding')
-                        ?? 'Model "%s" answered but returned no embedding vector.',
-                    $model->getName(),
-                ),
-            ))->jsonSerialize());
-        }
-
-        return new JsonResponse((new TestConnectionResponse(
-            success: true,
-            message: sprintf(
-                $this->translate('model.test.embedded')
-                    ?? 'Model "%s" returned a %d-dimension embedding (tokens: %d in).',
-                $model->getName(),
-                $dimensions,
-                $response->usage->promptTokens,
-            ),
-        ))->jsonSerialize());
-    }
 }
