+++ Tests/Unit/Controller/Backend/ModelControllerTest.php
@@ -11,20 +11,10 @@ namespace Netresearch\NrLlm\Tests\Unit\Controller\Backend;
 
 use LogicException;
 use Netresearch\NrLlm\Controller\Backend\ModelController;
-use Netresearch\NrLlm\Domain\Model\CompletionResponse;
-use Netresearch\NrLlm\Domain\Model\EmbeddingResponse;
 use Netresearch\NrLlm\Domain\Model\Model;
-use Netresearch\NrLlm\Domain\Model\Provider;
-use Netresearch\NrLlm\Domain\Model\UsageStatistics;
 use Netresearch\NrLlm\Domain\Repository\ModelRepository;
 use Netresearch\NrLlm\Domain\Repository\ProviderRepository;
-use Netresearch\NrLlm\Provider\Contract\ProviderInterface;
-use Netresearch\NrLlm\Provider\ProviderAdapterRegistryInterface;
-use Netresearch\NrLlm\Service\SetupWizard\DTO\DiscoveredModel;
-use Netresearch\NrLlm\Service\SetupWizard\ModelDiscoveryInterface;
-use Netresearch\NrLlm\Service\TestPromptResolverInterface;
 use PHPUnit\Framework\Attributes\AllowMockObjectsWithoutExpectations;
-use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
@@ -49,9 +39,6 @@ final class ModelControllerTest extends TestCase
     private ModelRepository&MockObject $modelRepository;
     private ProviderRepository&MockObject $providerRepository;
     private PersistenceManagerInterface&MockObject $persistenceManager;
-    private ProviderAdapterRegistryInterface&MockObject $providerAdapterRegistry;
-    private ModelDiscoveryInterface&MockObject $modelDiscovery;
-    private TestPromptResolverInterface&MockObject $testPromptResolver;
     private ModelController $subject;
     private mixed $previousBeUser;
 
@@ -69,10 +56,6 @@ final class ModelControllerTest extends TestCase
         $this->modelRepository = $this->createMock(ModelRepository::class);
         $this->providerRepository = $this->createMock(ProviderRepository::class);
         $this->persistenceManager = $this->createMock(PersistenceManagerInterface::class);
-        $this->providerAdapterRegistry = $this->createMock(ProviderAdapterRegistryInterface::class);
-        $this->modelDiscovery = $this->createMock(ModelDiscoveryInterface::class);
-        $this->testPromptResolver = $this->createMock(TestPromptResolverInterface::class);
-        $this->testPromptResolver->method('resolve')->willReturn('Hello, test prompt');
 
         // Create controller using reflection to inject only required dependencies
         $this->subject = $this->createControllerWithDependencies();
@@ -101,9 +84,6 @@ final class ModelControllerTest extends TestCase
         $this->setPrivateProperty($controller, 'modelRepository', $this->modelRepository);
         $this->setPrivateProperty($controller, 'providerRepository', $this->providerRepository);
         $this->setPrivateProperty($controller, 'persistenceManager', $this->persistenceManager);
-        $this->setPrivateProperty($controller, 'providerAdapterRegistry', $this->providerAdapterRegistry);
-        $this->setPrivateProperty($controller, 'modelDiscovery', $this->modelDiscovery);
-        $this->setPrivateProperty($controller, 'testPromptResolver', $this->testPromptResolver);
         // REC #8b: typed catches log via LoggerInterface — NullLogger
         // for unit tests so the property is initialised.
         $this->setPrivateProperty($controller, 'logger', new NullLogger());
@@ -517,479 +497,6 @@ final class ModelControllerTest extends TestCase
         self::assertStringContainsString('See system log', $data['error']);
     }
 
-    // testModelAction tests
-
-    #[Test]
-    public function testModelActionReturnsErrorForMissingUid(): void
-    {
-        $this->modelRepository
-            ->expects(self::never())
-            ->method('findByUid');
-
-        $request = $this->createRequest([]);
-        $response = $this->subject->testModelAction($request);
-
-        $data = $this->decodeJsonResponse($response);
-
-        self::assertSame(400, $response->getStatusCode());
-        self::assertArrayHasKey('error', $data);
-        self::assertIsString($data['error']);
-        self::assertStringContainsString('No model UID', $data['error']);
-    }
-
-    #[Test]
-    public function testModelActionReturnsErrorForNonexistentModel(): void
-    {
-        $this->modelRepository
-            ->expects(self::once())
-            ->method('findByUid')
-            ->with(99999)
-            ->willReturn(null);
-
-        $request = $this->createRequest(['uid' => 99999]);
-        $response = $this->subject->testModelAction($request);
-
-        $data = $this->decodeJsonResponse($response);
-
-        self::assertSame(404, $response->getStatusCode());
-        self::assertArrayHasKey('error', $data);
-        self::assertIsString($data['error']);
-        self::assertStringContainsString('not found', $data['error']);
-    }
-
-    #[Test]
-    public function testModelActionReturnsErrorForModelWithoutProvider(): void
-    {
-        $model = $this->createModel(1, true);
-        // Model has no provider set (null)
-
-        $this->modelRepository
-            ->expects(self::once())
-            ->method('findByUid')
-            ->with(1)
-            ->willReturn($model);
-
-        $request = $this->createRequest(['uid' => 1]);
-        $response = $this->subject->testModelAction($request);
-
-        $data = $this->decodeJsonResponse($response);
-
-        self::assertSame(400, $response->getStatusCode());
-        self::assertArrayHasKey('error', $data);
-        self::assertIsString($data['error']);
-        self::assertStringContainsString('no provider', $data['error']);
-    }
-
-    #[Test]
-    public function testModelActionReturnsSuccessOnSuccessfulTest(): void
-    {
-        $model = $this->createModel(1, true);
-        $model->setName('GPT-4');
-        $model->setModelId('gpt-4');
-
-        $provider = new Provider();
-        $providerReflection = new ReflectionClass($provider);
-        $providerUidProp = $providerReflection->getProperty('uid');
-        $providerUidProp->setValue($provider, 1);
-        $provider->setName('OpenAI');
-        $provider->setAdapterType('openai');
-        $model->setProvider($provider);
-
-        $this->modelRepository
-            ->expects(self::once())
-            ->method('findByUid')
-            ->with(1)
-            ->willReturn($model);
-
-        $adapter = $this->createMock(ProviderInterface::class);
-        $usage = new UsageStatistics(10, 5, 15);
-        $completionResponse = new CompletionResponse(
-            content: 'OK',
-            model: 'gpt-4',
-            usage: $usage,
-        );
-
-        $adapter
-            ->expects(self::once())
-            ->method('complete')
-            ->willReturn($completionResponse);
-
-        $this->providerAdapterRegistry
-            ->expects(self::once())
-            ->method('createAdapterFromModel')
-            ->with($model)
-            ->willReturn($adapter);
-
-        $request = $this->createRequest(['uid' => 1]);
-        $response = $this->subject->testModelAction($request);
-
-        $data = $this->decodeJsonResponse($response);
-
-        self::assertSame(200, $response->getStatusCode());
-        self::assertTrue($data['success']);
-        self::assertIsString($data['message']);
-        self::assertStringContainsString('GPT-4', $data['message']);
-        self::assertStringContainsString('OK', $data['message']);
-    }
-
-    #[Test]
-    public function testModelActionReturnsErrorOnAdapterException(): void
-    {
-        $model = $this->createModel(1, true);
-        $model->setName('GPT-4');
-        $model->setModelId('gpt-4');
-
-        $provider = new Provider();
-        $providerReflection = new ReflectionClass($provider);
-        $providerUidProp = $providerReflection->getProperty('uid');
-        $providerUidProp->setValue($provider, 1);
-        $provider->setName('OpenAI');
-        $provider->setAdapterType('openai');
-        $model->setProvider($provider);
-
-        $this->modelRepository
-            ->expects(self::once())
-            ->method('findByUid')
-            ->with(1)
-            ->willReturn($model);
-
-        $adapter = $this->createMock(ProviderInterface::class);
-        $adapter
-            ->method('complete')
-            ->willThrowException(new LogicException('API connection failed'));
-
-        $this->providerAdapterRegistry
-            ->expects(self::once())
-            ->method('createAdapterFromModel')
-            ->with($model)
-            ->willReturn($adapter);
-
-        $request = $this->createRequest(['uid' => 1]);
-        $response = $this->subject->testModelAction($request);
-
-        $data = $this->decodeJsonResponse($response);
-
-        self::assertSame(200, $response->getStatusCode());
-        self::assertFalse($data['success']);
-        self::assertIsString($data['message']);
-        self::assertStringContainsString('See system log', $data['message']);
-    }
-
-    private function createProvider(int $uid): Provider
-    {
-        $provider = new Provider();
-        $reflection = new ReflectionClass($provider);
-        $uidProperty = $reflection->getProperty('uid');
-        $uidProperty->setValue($provider, $uid);
-        $provider->setName('Test Provider');
-        $provider->setAdapterType('openai');
-        $provider->setApiKey('0190a5e0-7a1c-7b2d-8f3e-4a5b6c7d8e9f');
-        return $provider;
-    }
-
-    // fetchAvailableModelsAction tests
-
-    #[Test]
-    public function fetchAvailableModelsActionReturnsErrorForMissingProviderUid(): void
-    {
-        $this->providerRepository
-            ->expects(self::never())
-            ->method('findByUid');
-
-        $request = $this->createRequest([]);
-        $response = $this->subject->fetchAvailableModelsAction($request);
-
-        $data = $this->decodeJsonResponse($response);
-
-        self::assertSame(400, $response->getStatusCode());
-        self::assertArrayHasKey('error', $data);
-        self::assertIsString($data['error']);
-        self::assertStringContainsString('No provider UID', $data['error']);
-    }
-
-    #[Test]
-    public function fetchAvailableModelsActionReturnsErrorForNonexistentProvider(): void
-    {
-        $this->providerRepository
-            ->expects(self::once())
-            ->method('findByUid')
-            ->with(99999)
-            ->willReturn(null);
-
-        $request = $this->createRequest(['providerUid' => 99999]);
-        $response = $this->subject->fetchAvailableModelsAction($request);
-
-        $data = $this->decodeJsonResponse($response);
-
-        self::assertSame(404, $response->getStatusCode());
-        self::assertArrayHasKey('error', $data);
-        self::assertIsString($data['error']);
-        self::assertStringContainsString('Provider not found', $data['error']);
-    }
-
-    #[Test]
-    public function fetchAvailableModelsActionReturnsModelsFromProvider(): void
-    {
-        $provider = $this->createProvider(1);
-
-        $this->providerRepository
-            ->expects(self::once())
-            ->method('findByUid')
-            ->with(1)
-            ->willReturn($provider);
-
-        $discoveredModels = [
-            new DiscoveredModel(
-                modelId: 'gpt-4o',
-                name: 'GPT-4o',
-                description: 'Flagship model',
-                capabilities: ['chat', 'vision', 'tools'],
-                contextLength: 128000,
-                maxOutputTokens: 16384,
-            ),
-            new DiscoveredModel(
-                modelId: 'gpt-4o-mini',
-                name: 'GPT-4o Mini',
-                description: 'Fast and cheap',
-                capabilities: ['chat', 'tools'],
-                contextLength: 128000,
-                maxOutputTokens: 16384,
-            ),
-        ];
-
-        $this->modelDiscovery
-            ->expects(self::once())
-            ->method('discover')
-            ->willReturn($discoveredModels);
-        $this->modelDiscovery
-            ->method('wasLastDiscoveryFromFallback')
-            ->willReturn(false);
-
-        $request = $this->createRequest(['providerUid' => 1]);
-        $response = $this->subject->fetchAvailableModelsAction($request);
-
-        $data = $this->decodeJsonResponse($response);
-
-        self::assertSame(200, $response->getStatusCode());
-        self::assertTrue($data['success']);
-        self::assertSame('live', $data['source']);
-        self::assertIsArray($data['models']);
-        self::assertCount(2, $data['models']);
-        self::assertIsArray($data['models'][0]);
-        self::assertSame('gpt-4o', $data['models'][0]['id']);
-        self::assertSame('GPT-4o', $data['models'][0]['name']);
-        self::assertSame(128000, $data['models'][0]['contextLength']);
-        self::assertIsArray($data['models'][0]['capabilities']);
-        self::assertContains('vision', $data['models'][0]['capabilities']);
-    }
-
-    #[Test]
-    public function fetchAvailableModelsActionMarksFallbackSource(): void
-    {
-        $provider = $this->createProvider(1);
-
-        $this->providerRepository
-            ->expects(self::once())
-            ->method('findByUid')
-            ->with(1)
-            ->willReturn($provider);
-
-        $this->modelDiscovery
-            ->expects(self::once())
-            ->method('discover')
-            ->willReturn([
-                new DiscoveredModel(modelId: 'gpt-5.5', name: 'GPT-5.5'),
-            ]);
-        $this->modelDiscovery
-            ->method('wasLastDiscoveryFromFallback')
-            ->willReturn(true);
-
-        $request = $this->createRequest(['providerUid' => 1]);
-        $response = $this->subject->fetchAvailableModelsAction($request);
-
-        $data = $this->decodeJsonResponse($response);
-
-        self::assertSame(200, $response->getStatusCode());
-        self::assertTrue($data['success']);
-        self::assertSame('fallback', $data['source']);
-    }
-
-    #[Test]
-    public function fetchAvailableModelsActionReturnsErrorOnDiscoveryException(): void
-    {
-        $provider = $this->createProvider(1);
-
-        $this->providerRepository
-            ->expects(self::once())
-            ->method('findByUid')
-            ->with(1)
-            ->willReturn($provider);
-
-        $this->modelDiscovery
-            ->method('discover')
-            ->willThrowException(new LogicException('API unavailable'));
-
-        $request = $this->createRequest(['providerUid' => 1]);
-        $response = $this->subject->fetchAvailableModelsAction($request);
-
-        $data = $this->decodeJsonResponse($response);
-
-        self::assertSame(500, $response->getStatusCode());
-        self::assertFalse($data['success']);
-        self::assertIsString($data['error']);
-        self::assertStringContainsString('See system log', $data['error']);
-    }
-
-    // detectLimitsAction tests
-
-    #[Test]
-    public function detectLimitsActionReturnsErrorForMissingProviderUid(): void
-    {
-        $request = $this->createRequest(['modelId' => 'gpt-4o']);
-        $response = $this->subject->detectLimitsAction($request);
-
-        $data = $this->decodeJsonResponse($response);
-
-        self::assertSame(400, $response->getStatusCode());
-        self::assertArrayHasKey('error', $data);
-        self::assertIsString($data['error']);
-        self::assertStringContainsString('No provider UID', $data['error']);
-    }
-
-    #[Test]
-    public function detectLimitsActionReturnsErrorForMissingModelId(): void
-    {
-        $request = $this->createRequest(['providerUid' => 1]);
-        $response = $this->subject->detectLimitsAction($request);
-
-        $data = $this->decodeJsonResponse($response);
-
-        self::assertSame(400, $response->getStatusCode());
-        self::assertArrayHasKey('error', $data);
-        self::assertIsString($data['error']);
-        self::assertStringContainsString('No model ID', $data['error']);
-    }
-
-    #[Test]
-    public function detectLimitsActionReturnsErrorForNonexistentProvider(): void
-    {
-        $this->providerRepository
-            ->expects(self::once())
-            ->method('findByUid')
-            ->with(99999)
-            ->willReturn(null);
-
-        $request = $this->createRequest(['providerUid' => 99999, 'modelId' => 'gpt-4o']);
-        $response = $this->subject->detectLimitsAction($request);
-
-        $data = $this->decodeJsonResponse($response);
-
-        self::assertSame(404, $response->getStatusCode());
-        self::assertArrayHasKey('error', $data);
-        self::assertIsString($data['error']);
-        self::assertStringContainsString('Provider not found', $data['error']);
-    }
-
-    #[Test]
-    public function detectLimitsActionReturnsModelLimits(): void
-    {
-        $provider = $this->createProvider(1);
-
-        $this->providerRepository
-            ->expects(self::once())
-            ->method('findByUid')
-            ->with(1)
-            ->willReturn($provider);
-
-        $discoveredModels = [
-            new DiscoveredModel(
-                modelId: 'gpt-4o',
-                name: 'GPT-4o',
-                description: 'Flagship model',
-                capabilities: ['chat', 'vision', 'tools'],
-                contextLength: 128000,
-                maxOutputTokens: 16384,
-            ),
-        ];
-
-        $this->modelDiscovery
-            ->expects(self::once())
-            ->method('discover')
-            ->willReturn($discoveredModels);
-
-        $request = $this->createRequest(['providerUid' => 1, 'modelId' => 'gpt-4o']);
-        $response = $this->subject->detectLimitsAction($request);
-
-        $data = $this->decodeJsonResponse($response);
-
-        self::assertSame(200, $response->getStatusCode());
-        self::assertTrue($data['success']);
-        self::assertSame(128000, $data['contextLength']);
-        self::assertSame(16384, $data['maxOutputTokens']);
-        self::assertIsArray($data['capabilities']);
-        self::assertContains('vision', $data['capabilities']);
-    }
-
-    #[Test]
-    public function detectLimitsActionReturnsErrorWhenModelNotFound(): void
-    {
-        $provider = $this->createProvider(1);
-
-        $this->providerRepository
-            ->expects(self::once())
-            ->method('findByUid')
-            ->with(1)
-            ->willReturn($provider);
-
-        $discoveredModels = [
-            new DiscoveredModel(
-                modelId: 'gpt-4o-mini',
-                name: 'GPT-4o Mini',
-            ),
-        ];
-
-        $this->modelDiscovery
-            ->expects(self::once())
-            ->method('discover')
-            ->willReturn($discoveredModels);
-
-        $request = $this->createRequest(['providerUid' => 1, 'modelId' => 'gpt-4o']);
-        $response = $this->subject->detectLimitsAction($request);
-
-        $data = $this->decodeJsonResponse($response);
-
-        self::assertSame(404, $response->getStatusCode());
-        self::assertFalse($data['success']);
-        self::assertIsString($data['error']);
-        self::assertStringContainsString('not found', $data['error']);
-    }
-
-    #[Test]
-    public function detectLimitsActionReturnsErrorOnDiscoveryException(): void
-    {
-        $provider = $this->createProvider(1);
-
-        $this->providerRepository
-            ->expects(self::once())
-            ->method('findByUid')
-            ->with(1)
-            ->willReturn($provider);
-
-        $this->modelDiscovery
-            ->method('discover')
-            ->willThrowException(new LogicException('API unavailable'));
-
-        $request = $this->createRequest(['providerUid' => 1, 'modelId' => 'gpt-4o']);
-        $response = $this->subject->detectLimitsAction($request);
-
-        $data = $this->decodeJsonResponse($response);
-
-        self::assertSame(500, $response->getStatusCode());
-        self::assertFalse($data['success']);
-        self::assertIsString($data['error']);
-        self::assertStringContainsString('See system log', $data['error']);
-    }
-
     // Edge case tests for non-array body
 
     #[Test]
@@ -1035,193 +542,5 @@ final class ModelControllerTest extends TestCase
 
         self::assertSame(400, $response->getStatusCode());
         self::assertArrayHasKey('error', $data);
-    }
-
-    #[Test]
-    public function testModelActionWithNonArrayBodyReturnsError(): void
-    {
-        $request = (new ServerRequest('/ajax/test', 'POST'))
-            // @phpstan-ignore-next-line Intentionally passing invalid type to test error handling
-            ->withParsedBody('not an array');
-
-        $response = $this->subject->testModelAction($request);
-
-        $data = $this->decodeJsonResponse($response);
-
-        self::assertSame(400, $response->getStatusCode());
-        self::assertArrayHasKey('error', $data);
-    }
-
-    #[Test]
-    public function fetchAvailableModelsActionWithNonArrayBodyReturnsError(): void
-    {
-        $request = (new ServerRequest('/ajax/test', 'POST'))
-            // @phpstan-ignore-next-line Intentionally passing invalid type to test error handling
-            ->withParsedBody('not an array');
-
-        $response = $this->subject->fetchAvailableModelsAction($request);
-
-        $data = $this->decodeJsonResponse($response);
-
-        self::assertSame(400, $response->getStatusCode());
-        self::assertArrayHasKey('error', $data);
-    }
-
-    #[Test]
-    public function detectLimitsActionWithNonArrayBodyReturnsError(): void
-    {
-        $request = (new ServerRequest('/ajax/test', 'POST'))
-            // @phpstan-ignore-next-line Intentionally passing invalid type to test error handling
-            ->withParsedBody('not an array');
-
-        $response = $this->subject->detectLimitsAction($request);
-
-        $data = $this->decodeJsonResponse($response);
-
-        self::assertSame(400, $response->getStatusCode());
-        self::assertArrayHasKey('error', $data);
-    }
-
-    #[Test]
-    public function anImageOnlyModelIsNotSentAChatPrompt(): void
-    {
-        $provider = new Provider();
-        $provider->setName('OpenAI');
-
-        $model = new Model();
-        $model->setName('dall-e-3');
-        $model->setProvider($provider);
-        $model->setCapabilities('image');
-
-        $this->modelRepository->method('findByUid')->willReturn($model);
-        // The whole point: no adapter is built, so nothing is sent upstream.
-        $this->providerAdapterRegistry->expects(self::never())->method('createAdapterFromModel');
-
-        $body = $this->decodeJsonResponse($this->subject->testModelAction($this->createRequest(['uid' => 7])));
-
-        self::assertFalse($body['success']);
-        self::assertIsString($body['message']);
-        self::assertStringContainsString('dall-e-3', $body['message']);
-        self::assertStringContainsString('Extension Configuration', $body['message']);
-    }
-
-    #[Test]
-    public function anEmbeddingModelIsProbedWithAnEmbeddingCall(): void
-    {
-        $provider = new Provider();
-        $provider->setName('OpenAI');
-
-        $model = new Model();
-        $model->setName('text-embedding-3-small');
-        $model->setProvider($provider);
-        $model->setCapabilities('embeddings');
-
-        $this->modelRepository->method('findByUid')->willReturn($model);
-
-        $adapter = $this->createMock(ProviderInterface::class);
-        $adapter->expects(self::once())->method('embeddings')->willReturn(
-            new EmbeddingResponse([[0.1, 0.2, 0.3]], 'text-embedding-3-small', new UsageStatistics(4, 0, 4)),
-        );
-        $adapter->expects(self::never())->method('complete');
-        $this->providerAdapterRegistry->method('createAdapterFromModel')->willReturn($adapter);
-
-        $body = $this->decodeJsonResponse($this->subject->testModelAction($this->createRequest(['uid' => 8])));
-
-        self::assertTrue($body['success']);
-        self::assertIsString($body['message']);
-        self::assertStringContainsString('3-dimension', $body['message']);
-    }
-
-    #[Test]
-    public function aModelWithoutDeclaredCapabilitiesKeepsTheChatProbe(): void
-    {
-        $provider = new Provider();
-        $provider->setName('OpenAI');
-
-        $model = new Model();
-        $model->setName('unlabelled');
-        $model->setProvider($provider);
-        // Capabilities are optional in TCA and routinely left empty; refusing
-        // to test that case would be a regression.
-
-        $this->modelRepository->method('findByUid')->willReturn($model);
-
-        $adapter = $this->createMock(ProviderInterface::class);
-        $adapter->expects(self::once())->method('complete')->willReturn(
-            new CompletionResponse('hi', 'unlabelled', new UsageStatistics(3, 1, 4)),
-        );
-        $this->providerAdapterRegistry->method('createAdapterFromModel')->willReturn($adapter);
-
-        $body = $this->decodeJsonResponse($this->subject->testModelAction($this->createRequest(['uid' => 9])));
-
-        self::assertTrue($body['success']);
-    }
-
-    /**
-     * @return array<string, array{string}>
-     */
-    public static function chatShapedCapabilityProvider(): array
-    {
-        return [
-            'chat' => ['chat'],
-            'completion' => ['completion'],
-            'vision' => ['vision'],
-            'tools' => ['tools'],
-            'streaming' => ['streaming'],
-            'json_mode' => ['json_mode'],
-            // Audio is a chat-completions modality here — there is no audio
-            // service and no separate credential.
-            'audio' => ['audio'],
-        ];
-    }
-
-    #[Test]
-    #[DataProvider('chatShapedCapabilityProvider')]
-    public function everyChatShapedCapabilityKeepsTheChatProbe(string $capability): void
-    {
-        $provider = new Provider();
-        $provider->setName('OpenAI');
-
-        $model = new Model();
-        $model->setName('some-model');
-        $model->setProvider($provider);
-        $model->setCapabilities($capability);
-
-        $this->modelRepository->method('findByUid')->willReturn($model);
-
-        $adapter = $this->createMock(ProviderInterface::class);
-        $adapter->expects(self::once())->method('complete')->willReturn(
-            new CompletionResponse('hi', 'some-model', new UsageStatistics(3, 1, 4)),
-        );
-        $this->providerAdapterRegistry->method('createAdapterFromModel')->willReturn($adapter);
-
-        $body = $this->decodeJsonResponse($this->subject->testModelAction($this->createRequest(['uid' => 11])));
-
-        self::assertTrue($body['success']);
-    }
-
-    #[Test]
-    public function anEmbeddingProbeThatReturnsNoVectorIsNotASuccess(): void
-    {
-        $provider = new Provider();
-        $provider->setName('OpenAI');
-
-        $model = new Model();
-        $model->setName('text-embedding-3-small');
-        $model->setProvider($provider);
-        $model->setCapabilities('embeddings');
-
-        $this->modelRepository->method('findByUid')->willReturn($model);
-
-        $adapter = $this->createMock(ProviderInterface::class);
-        // A 2xx whose body carries no vector: nothing was verified.
-        $adapter->method('embeddings')->willReturn(
-            new EmbeddingResponse([], 'text-embedding-3-small', new UsageStatistics(4, 0, 4)),
-        );
-        $this->providerAdapterRegistry->method('createAdapterFromModel')->willReturn($adapter);
-
-        $body = $this->decodeJsonResponse($this->subject->testModelAction($this->createRequest(['uid' => 12])));
-
-        self::assertFalse($body['success']);
     }
 }
